chore(agent-gui): satisfy fork checks
